+++ app/Http/Controllers/CarrerasFacultadController.php
@@ -17,14 +17,26 @@ class CarrerasFacultadController extends Controller
      */
     public function index()
     {
+        $logs = new LogsController();
+        $sesion = new SesionController();
+        $sesion->setTipoUsuario($logs->getLogInfo());
+        if($sesion->getTipoUsuario()=='Administrador'){
+            #Acciones de Admin
+            $carrerasFacultad = \DB::table('CarreraFacultad')
+            ->join('Carrera','CarreraFacultad.idcarrera','=','Carrera.idcarrera')
+            ->join('Facultad','CarreraFacultad.idfacultad','=','Facultad.idfacultad')
+            ->join('TipoCarrera','Carrera.tipocarrera','=','TipoCarrera.idtipocarrera')
+            ->select('Carrera.carrera','Facultad.facultad','TipoCarrera.tipocarrera')
+            ->get();
+            return $carrerasFacultad;
+        }else if($sesion->getTipoUsuario()=='Egresado'){
+            return response()->json(['Mensaje'=>'Usuario no autorizado'],401);
+            #Acciones de Egresado
+        }else{
+            #No loggeado/No se reconoce sesión
+            return response()->json(['Mensaje'=>'Sesión no iniciada'],403);
+        }
         //
-        $carrerasFacultad = \DB::table('CarreraFacultad')
-        ->join('Carrera','CarreraFacultad.idcarrera','=','Carrera.idcarrera')
-        ->join('Facultad','CarreraFacultad.idfacultad','=','Facultad.idfacultad')
-        ->join('TipoCarrera','Carrera.tipocarrera','=','TipoCarrera.idtipocarrera')
-        ->select('Carrera.carrera','Facultad.facultad','TipoCarrera.tipocarrera')
-        ->get();
-        return $carrerasFacultad;
     }
 
     /**
@@ -45,17 +57,29 @@ class CarrerasFacultadController extends Controller
      */
     public function store(Request $request)
     {
-        $carreraFacultad = new CarreraFacultad;
-        $carrera = new Carrera;
-
-        $carrera->carrera = $request->carrera;
-        $carrera->tipocarrera = $request->tipocarrera;
-        $carrera->save();
-        $carreraFacultad->idcarrera = \DB::table('Carrera')->latest('idcarrera')->first()->idcarrera;
-        $carreraFacultad->idfacultad = $request->idfacultad;
-        $carreraFacultad->save();
-
-        return response()->json(['Mensaje'=>'Carrera agregada exitosamente'],200);
+        $logs = new LogsController();
+        $sesion = new SesionController();
+        $sesion->setTipoUsuario($logs->getLogInfo());
+        if($sesion->getTipoUsuario()=='Administrador'){
+            #Acciones de Admin
+            $carreraFacultad = new CarreraFacultad;
+            $carrera = new Carrera;
+    
+            $carrera->carrera = $request->carrera;
+            $carrera->getTipocarrera = $request->getTipocarrera;
+            $carrera->save();
+            $carreraFacultad->idcarrera = \DB::table('Carrera')->latest('idcarrera')->first()->idcarrera;
+            $carreraFacultad->idfacultad = $request->idfacultad;
+            $carreraFacultad->save();
+    
+            return response()->json(['Mensaje'=>'Carrera agregada exitosamente'],200);
+        }else if($sesion->getTipoUsuario()=='Egresado'){
+            #Acciones de Egresado
+            return response()->json(['Mensaje' =>'Usuario no autorizado'],401);
+        }else{
+            #No loggeado/No se reconoce sesión
+            return response()->json(['Mensaje'=>'Sesión no iniciada'],403);
+        }
         //
         //
     }
@@ -68,18 +92,29 @@ class CarrerasFacultadController extends Controller
      */
     public function show($carrera)
     {
-
-        $carreras = \DB::table('Carrera')
-        ->where('Carrera.idcarrera',$carrera)
-        ->join('CarreraFacultad','CarreraFacultad.idcarrera','=','Carrera.idcarrera')
-        ->join('Facultad','CarreraFacultad.idfacultad','=','Facultad.idfacultad')
-        ->select('Carrera.carrera','Facultad.facultad')
-        ->get();
-
-        if(!$carreras){
-            return response()->json(['mensaje'=>'Carrera inexistente']);
+        $logs = new LogsController();
+        $sesion = new SesionController();
+        $sesion->setTipoUsuario($logs->getLogInfo());
+        if($sesion->getTipoUsuario()=='Administrador'){
+            #Acciones de Admin
+            $carreras = \DB::table('Carrera')
+            ->where('Carrera.idcarrera',$carrera)
+            ->join('CarreraFacultad','CarreraFacultad.idcarrera','=','Carrera.idcarrera')
+            ->join('Facultad','CarreraFacultad.idfacultad','=','Facultad.idfacultad')
+            ->select('Carrera.carrera','Facultad.facultad')
+            ->get();
+    
+            if(!$carreras){
+                return response()->json(['mensaje'=>'Carrera inexistente']);
+            }else{
+                return $carreras;
+            }
+        }else if($sesion->getTipoUsuario()=='Egresado'){
+            #Acciones de Egresado
+            return response()->json(['Mensaje' =>'Usuario no autorizado'],401);
         }else{
-            return $carreras;
+            #No loggeado/No se reconoce sesión
+            return response()->json(['Mensaje'=>'Sesión no iniciada'],403);
         }
         //
     }
@@ -104,10 +139,22 @@ class CarrerasFacultadController extends Controller
      */
     public function update(Request $request, $idcarrera)
     {
-        $carrera = Carrera::find($idcarrera);
-        $carrera->carrera = $request->carrera;
-        $carrera->save();
-        return response()->json(['Mensaje'=>'Carrera modificada exitosamente'],200);
+        $logs = new LogsController();
+        $sesion = new SesionController();
+        $sesion->setTipoUsuario($logs->getLogInfo());
+        if($sesion->getTipoUsuario()=='Administrador'){
+            #Acciones de Admin
+            $carrera = Carrera::find($idcarrera);
+            $carrera->carrera = $request->carrera;
+            $carrera->save();
+            return response()->json(['Mensaje'=>'Carrera modificada exitosamente'],200);
+        }else if($sesion->getTipoUsuario()=='Egresado'){
+            #Acciones de Egresado
+            return response()->json(['Mensaje' =>'Usuario no autorizado'],401);
+        }else{
+            #No loggeado/No se reconoce sesión
+            return response()->json(['Mensaje'=>'Sesión no iniciada'],403);
+        }
         //
     }
 
@@ -119,9 +166,21 @@ class CarrerasFacultadController extends Controller
      */
     public function destroy($idcarrera)
     {
-        $carrera = Carrera::find($idcarrera);
-        $carrera->delete();
-        return response()->json(['Mensaje'=>'Elemento eliminardo'],200);
+        $logs = new LogsController();
+        $sesion = new SesionController();
+        $sesion->setTipoUsuario($logs->getLogInfo());
+        if($sesion->getTipoUsuario()=='Administrador'){
+            #Acciones de Admin
+            $carrera = Carrera::find($idcarrera);
+            $carrera->delete();
+            return response()->json(['Mensaje'=>'Elemento eliminardo'],200);
+        }else if($sesion->getTipoUsuario()=='Egresado'){
+            #Acciones de Egresado
+            return response()->json(['Mensaje' =>'Usuario no autorizado'],401);
+        }else{
+            #No loggeado/No se reconoce sesión
+            return response()->json(['Mensaje'=>'Sesión no iniciada'],403);
+        }
         //
     }
 }
